# control-bridge/openapi-gpt-actions.yaml
openapi: 3.0.0
info:
  title: Control Bridge API
  description: |
    GPT Actions bridge for marbomebel.ru.
    Jobs are auto-dispatched on creation — NO confirmation needed.
    Flow: createJob → poll getJobStatus → getJobResult when completed.
    Only ONE job can run at a time. Call getActiveJobs first to check.
  version: "1.4.1"
servers:
  - url: https://api.marbomebel.ru
    description: Production

paths:
  /health:
    get:
      operationId: healthCheck
      summary: "Health check"
      responses:
        "200":
          description: "Service healthy"
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok

  /jobs/active:
    get:
      operationId: getActiveJobs
      summary: "Check if any jobs are currently running. ALWAYS call this BEFORE createJob. If non-empty, do NOT create a new job — tell the user to wait."
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "List of active jobs. Empty array means safe to create a new job."
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    job_id:
                      type: string
                    status:
                      type: string
                    title:
                      type: string
                    created_at:
                      type: number

  /jobs:
    post:
      operationId: createJob
      summary: "Create and auto-start a new job. No confirmation needed — it starts immediately. Returns 429 if another job is running."
      description: |
        Creates a job and automatically dispatches it to Claude Code.
        DO NOT call confirmJob after this — it is not needed.
        Just poll getJobStatus until status is 'completed' or 'failed'.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: "Short title for the job"
                raw_user_request:
                  type: string
                  description: "Original user request text"
                normalized_brief:
                  type: string
                  description: "Normalized brief for execution"
                  default: ""
              required:
                - title
                - raw_user_request
      responses:
        "200":
          description: "Job created and auto-dispatched"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
        "429":
          description: "Another job is already running. Tell the user to wait and try later."

  /jobs/{job_id}/status:
    get:
      operationId: getJobStatus
      summary: "Poll job status. Returns queued/dispatched/running/completed/failed/cancelled."
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Current job status"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
        "404":
          description: "Job not found"

  /jobs/{job_id}/result:
    get:
      operationId: getJobResult
      summary: "Get job result (only meaningful when status is completed)"
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Job result"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResultResponse"
        "404":
          description: "Job not found"

  /jobs/{job_id}/cancel:
    post:
      operationId: cancelJob
      summary: "Cancel a running or queued job"
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Job cancelled"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
        "404":
          description: "Job not found"
        "409":
          description: "Job cannot be cancelled in current state"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

  schemas:
    JobResponse:
      type: object
      properties:
        job_id:
          type: string
        status:
          type: string
          enum:
            - queued
            - dispatched
            - running
            - completed
            - failed
            - cancelled
      required:
        - job_id
        - status

    JobResultData:
      type: object
      properties:
        summary:
          type: string
          description: "Human-readable summary of what was done"
        changes_made:
          type: array
          items:
            type: string
        artifacts:
          type: array
          items:
            type: string
        tests_run:
          type: array
          items:
            type: string
        manual_checks:
          type: array
          items:
            type: string
        risks_remaining:
          type: array
          items:
            type: string
      required:
        - summary

    JobResultResponse:
      type: object
      properties:
        job_id:
          type: string
        status:
          type: string
        result:
          $ref: "#/components/schemas/JobResultData"
      required:
        - job_id
        - status